+++ src/Login.js
@@ -1,3 +1,35 @@
+/**
+ * Renders user login dialog
+ * 
+ * @link    https://airmencoders.cce.us.af.mil/mqf
+ * @link    https://github.com/airmencoders/mqfeditor
+ * @file    Login.js
+ * @author  chris-m92
+ * @since   0.1.0
+ * @version 0.6.0
+ * 
+ * MIT License
+ * 
+ * Copyright (c) 2020 Airmen Coders
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
 import React from 'react'
 
 import Button from '@material-ui/core/Button'
@@ -26,9 +58,13 @@ const useStyles = makeStyles((theme) => ({
     flex: '0 0 100%',
     padding: theme.spacing(3),
   },
+  cacImage: {
+    border: '2px solid gray',
+    borderRadius: 5,
+  },
 }))
 
-const Login = ({ state, onClick }) => {
+const Login = ({ state, onLoginClick }) => {
   const classes = useStyles()
 
   return (
@@ -46,12 +82,12 @@ const Login = ({ state, onClick }) => {
             <CardActions>
               <Grid container direction="row" alignItems="flex-end" justify="space-between">
                 <Grid item xs={3}>
-                  <CardMedia component="img" alt="CAC" src={cac} title="CAC" />
+                  <CardMedia component="img" alt="CAC" src={cac} title="CAC" className={classes.cacImage} />
                 </Grid>
                 <Grid item>
                   <Button
                     variant="contained"
-                    onClick={onClick}
+                    onClick={onLoginClick}
                     color="primary"
                   >Log In with CAC</Button>
                 </Grid>