+++ src/ResponsiveNavigation.js
@@ -2,6 +2,9 @@ import React from 'react'
 import { NavLink } from 'react-router-dom'
 
 import AppBar from '@material-ui/core/AppBar'
+import Avatar from '@material-ui/core/Avatar'
+import Button from '@material-ui/core/Button'
+import Divider from '@material-ui/core/Divider'
 import IconButton from '@material-ui/core/IconButton'
 import Menu from '@material-ui/core/Menu'
 import MenuItem from '@material-ui/core/MenuItem'
@@ -9,14 +12,9 @@ import { makeStyles } from '@material-ui/core/styles'
 import Toolbar from '@material-ui/core/Toolbar'
 import Typography from '@material-ui/core/Typography'
 
-import AccountCircle from '@material-ui/icons/AccountCircle'
 import MenuIcon from '@material-ui/icons/Menu'
-import MoreIcon from '@material-ui/icons/MoreVert'
 
 const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-  },
   grow: {
     flexGrow: 1,
   },
@@ -29,85 +27,24 @@ const useStyles = makeStyles((theme) => ({
   title: {
     flexGrow: 1,
   },
-  sectionDesktop: {
-    display: 'none',
-    [theme.breakpoints.up('md')]: {
-      display: 'flex',
-    },
-  },
-  sectionMobile: {
-    display: 'flex',
-    [theme.breakpoints.up('md')]: {
-      display: 'none',
-    },
-  },
   appBar: {
     zIndex: theme.zIndex.drawer + 1,
   }
 }))
 
-const Navigation = ({ state, onMenuClick }) => {
+const Navigation = ({ state, onMenuClick, onLogoutClick }) => {
   const classes = useStyles()
   const [anchorEl, setAnchorEl] = React.useState(null)
-  const [mobileMoreAnchorEl, setMobileMoreAnchorEl] = React.useState(null)
 
-  const isMenuOpen = Boolean(anchorEl)
-  const isMobileMenuOpen = Boolean(mobileMoreAnchorEl)
+  const open = Boolean(anchorEl)
 
-  const handleProfileMenuOpen = (event) => {
+  const handleMenu = (event) => {
     setAnchorEl(event.currentTarget)
   }
 
-  const handleMobileMenuClose = () => {
-    setMobileMoreAnchorEl(null)
-  }
-
-  const handleMenuClose = () => {
+  const handleClose = () => {
     setAnchorEl(null)
-    handleMobileMenuClose()
   }
-
-  const handleMobileMenuOpen = (event) => {
-    setMobileMoreAnchorEl(event.currentTarget)
-  }
-
-  const menuId = 'primary-account-menu'
-  const renderMenu = (
-    <Menu
-      anchorEl={anchorEl}
-      anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-      id={menuId}
-      keepMounted
-      transformOrigin={{ vertical: 'top', horizontal: 'right' }}
-      open={isMenuOpen}
-      onClose={handleMenuClose}
-    >
-      <MenuItem onClick={handleMenuClose}>
-        <NavLink to={`/u/${state.user.id}`} style={{ textDecoration: 'none', color: 'initial' }}>
-          Account Settings
-        </NavLink>
-      </MenuItem>
-    </Menu>
-  )
-
-  const mobileMenuId = 'primary-account-menu-mobile'
-  const renderMobileMenu = (
-    <Menu
-      anchorEl={mobileMoreAnchorEl}
-      anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-      id={mobileMenuId}
-      keepMounted
-      transformOrigin={{ vertical: 'top', horizontal: 'right' }}
-      open={isMobileMenuOpen}
-      onClose={handleMobileMenuClose}
-    >
-      <MenuItem onClick={handleProfileMenuOpen}>
-        <NavLink to={`/u/${state.user.id}`} style={{ textDecoration: 'none', color: 'initial' }}>
-          Account Settings
-        </NavLink>
-      </MenuItem>
-    </Menu>
-  )
 
   return (
     <div className={classes.grow}>
@@ -131,39 +68,40 @@ const Navigation = ({ state, onMenuClick }) => {
           </div>
           <div className={classes.grow} />
           {state.isAuthenticated && (
-            <React.Fragment>
-              <div className={classes.sectionDesktop}>
-                <IconButton
-                  edge="end"
-                  aria-label="account of current user"
-                  aria-controls={menuId}
-                  aria-haspopup="true"
-                  onClick={handleProfileMenuOpen}
-                  color="inherit"
-                >
-                  <AccountCircle />
-                  <Typography variant="subtitle2">
-                    {(state.user) ? state.user.display : ""}
-                  </Typography>
-                </IconButton>
-              </div>
-              <div className={classes.sectionMobile}>
-                <IconButton
-                  aria-label="show more"
-                  aria-controls={mobileMenuId}
-                  aria-haspopup="true"
-                  onClick={handleMobileMenuOpen}
-                  color="inherit"
-                >
-                  <MoreIcon />
-                </IconButton>
-              </div>
-            </React.Fragment>
+            <div>
+              <Button
+                aria-label='user account'
+                aria-controls='account-menu'
+                aria-haspopup='true'
+                onClick={handleMenu}
+                color='inherit'  
+              >
+                <Avatar>{`${state.user.first.substring(0,1)}${state.user.last.substring(0,1)}`}</Avatar>
+              </Button>
+              <Menu
+                id='account-menu'
+                anchorEl={anchorEl}
+                anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
+                keepMounted
+                transformOrigin={{ vertical: 'top', horizontal: 'right' }}
+                open={open}
+                onClose={handleClose}
+              >
+                <MenuItem onClick={handleClose}>
+                  <NavLink to={`/u/${state.user.id}`} style={{ textDecoration: 'none', color: 'inherit' }} >
+                    Account Settings
+                  </NavLink>
+                </MenuItem>
+                <MenuItem onClick={onLogoutClick} >
+                    Log Out
+                </MenuItem>
+                <Divider />
+                v0.6.0
+              </Menu>
+            </div>
           )}
         </Toolbar>
       </AppBar>
-      {renderMobileMenu}
-      {renderMenu}
     </div>
   )
 }